+++ src/page/pageIndex.tsx
@@ -1,0 +1,4 @@
+export enum pageIndex{
+    uploadFile,
+    outputPreference
+}
